Extract model resource array utilities
Move array joint grouping/indexing helpers out of ModelResourceExporter behind the existing public static facade, with focused characterization tests for the preserved behavior.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
